docs: employ Rust's contributors-retain-copyright policy
As far as I know, Rust's COPYRIGHT file is licensed under the MIT
license (and the Apache 2.0 license) and would not be considered
a substantial part of `rustc`, so it's likely fine to "borrow" it like
this. If I'm wrong, please contact me. I didn't want to run the risk of
writing something that wouldn't hold up to legal scrutiny.