backend: fix if exprs with no else block when the condition is false
This fixes a bug where if expressions didn't return a value if the
condition was false and there was no `else` block. This meant that
an if statement like this:

if false {}

...would emit a pop to discard the return value, but not a push to
push a return value. This was fixed by attaching an empty `else` block
to any if expression that doesn't have one when lowering it.

In the future, `if` expressions with no `else` blocks will only be
allowed as statements, so always returning () should be fine.