vm: add some low-level functions for project triplicata
This enables it to use Laria as a scripting language long before it is
ready, though it uses several hacks to pull this off. I doubt this
would be useful in the main branch, at least for now. All of this will
probably be replaced by language features in the future, namely
coroutines.